changelist: generate changelist from uploaded gerrit commits
The purpose of this script is to search your local git repository and use the
change-id information and obtain information about uploaded git commits

This can be helpful when sharing the changes with other people. Obviously
this requires due diligence with the maintenance of Change-Id information in
the git commits.

There are also some useful utils like: changelist -G, which will inspect
the last four commits and try to find the project and branch of the commits.